# Conflicts:
#	all.gyp
#	third_party/gflags/gen/win/include/private/config.h
#	webrtc/BUILD.gn
#	webrtc/api/call/audio_receive_stream.h
#	webrtc/api/mediastreamtrackproxy.h
#	webrtc/api/peerconnection.cc
#	webrtc/api/peerconnection.h
#	webrtc/api/peerconnectionfactory.cc
#	webrtc/api/peerconnectionfactory.h
#	webrtc/api/peerconnectionfactoryproxy.h
#	webrtc/api/rtpreceiver.cc
#	webrtc/api/rtpreceiver.h
#	webrtc/api/statscollector.cc
#	webrtc/api/videotrack.h
#	webrtc/api/videotracksource.cc
#	webrtc/api/videotracksource.h
#	webrtc/audio/audio_receive_stream.cc
#	webrtc/base/BUILD.gn
#	webrtc/base/autodetectproxy.cc
#	webrtc/base/base.gyp
#	webrtc/base/bitbuffer_unittest.cc
#	webrtc/base/common.cc
#	webrtc/base/diskcache_win32.cc
#	webrtc/base/logging.cc
#	webrtc/base/openssladapter.cc
#	webrtc/base/proxydetect.cc
#	webrtc/base/socketadapters.cc
#	webrtc/base/systeminfo.cc
#	webrtc/base/task_queue.h
#	webrtc/base/win32filesystem.cc
#	webrtc/base/windowpickerfactory.h
#	webrtc/build/common.gypi
#	webrtc/build/gyp_webrtc.py
#	webrtc/common_audio/common_audio.gyp
#	webrtc/common_types.h
#	webrtc/common_video/common_video.gyp
#	webrtc/common_video/incoming_video_stream.cc
#	webrtc/common_video/video_render_frames.cc
#	webrtc/engine_configurations.h
#	webrtc/libjingle/xmpp/pingtask.cc
#	webrtc/media/base/videobroadcaster.h
#	webrtc/media/engine/webrtcvideocapturer.cc
#	webrtc/media/engine/webrtcvoiceengine.cc
#	webrtc/media/media.gyp
#	webrtc/media/sctp/sctpdataengine.cc
#	webrtc/modules/audio_coding/codecs/isac/isacfix.gypi
#	webrtc/modules/audio_coding/codecs/opus/opus.gypi
#	webrtc/modules/audio_coding/neteq/neteq.gypi
#	webrtc/modules/audio_coding/test/Tester.cc
#	webrtc/modules/audio_device/audio_device.gypi
#	webrtc/modules/audio_device/audio_device_impl.cc
#	webrtc/modules/audio_device/test/audio_device_test_api.cc
#	webrtc/modules/audio_processing/audio_processing.gypi
#	webrtc/modules/bitrate_controller/bitrate_controller.gypi
#	webrtc/modules/modules.gyp
#	webrtc/modules/remote_bitrate_estimator/remote_bitrate_estimator.gypi
#	webrtc/modules/video_capture/video_capture.gypi
#	webrtc/modules/video_capture/video_capture_impl.h
#	webrtc/modules/video_capture/windows/video_capture_factory_windows.cc
#	webrtc/modules/video_coding/codecs/h264/h264_encoder_impl.cc
#	webrtc/modules/video_coding/codecs/h264/h264_encoder_impl.h
#	webrtc/modules/video_coding/codecs/h264/h264_video_toolbox_encoder.h
#	webrtc/modules/video_coding/codecs/h264/h264_video_toolbox_encoder.mm
#	webrtc/modules/video_coding/codecs/i420/include/i420.h
#	webrtc/modules/video_coding/codecs/test/videoprocessor_integrationtest.cc
#	webrtc/modules/video_coding/codecs/vp8/simulcast_encoder_adapter.cc
#	webrtc/modules/video_coding/codecs/vp8/simulcast_encoder_adapter.h
#	webrtc/modules/video_coding/codecs/vp8/vp8_impl.cc
#	webrtc/modules/video_coding/codecs/vp8/vp8_impl.h
#	webrtc/modules/video_coding/codecs/vp9/vp9_impl.h
#	webrtc/modules/video_coding/generic_decoder.cc
#	webrtc/modules/video_coding/generic_encoder.cc
#	webrtc/modules/video_coding/timing.h
#	webrtc/modules/video_coding/utility/h264_bitstream_parser.cc
#	webrtc/modules/video_coding/video_coding_test.gypi
#	webrtc/modules/video_coding/video_receiver.cc
#	webrtc/modules/video_coding/video_sender.cc
#	webrtc/modules/video_processing/video_processing.gypi
#	webrtc/p2p/p2p.gyp
#	webrtc/supplement.gypi
#	webrtc/system_wrappers/system_wrappers.gyp
#	webrtc/test/channel_transport/udp_socket_manager_wrapper.cc
#	webrtc/test/channel_transport/udp_socket_wrapper.cc
#	webrtc/test/channel_transport/udp_transport_impl.cc
#	webrtc/test/run_all_unittests.cc
#	webrtc/test/test.gyp
#	webrtc/test/testsupport/fileutils_unittest.cc
#	webrtc/tools/tools.gyp
#	webrtc/video/payload_router_unittest.cc
#	webrtc/video/receive_statistics_proxy.cc
#	webrtc/video/receive_statistics_proxy.h
#	webrtc/video/video_encoder.cc
#	webrtc/video/video_encoder_unittest.cc
#	webrtc/video/video_stream_decoder.cc
#	webrtc/video_encoder.h
#	webrtc/voice_engine/channel.cc
#	webrtc/voice_engine/voe_hardware_impl.cc
#	webrtc/webrtc_examples.gyp
#	webrtc/webrtc_tests.gypi

+++ webrtc/test/fake_network_pipe_unittest.cc
@@ -10,13 +10,12 @@
 
 #include <memory>
 
-#include "testing/gmock/include/gmock/gmock.h"
-#include "testing/gtest/include/gtest/gtest.h"
-
-#include "webrtc/call.h"
+#include "webrtc/call/call.h"
 #include "webrtc/system_wrappers/include/clock.h"
 #include "webrtc/system_wrappers/include/tick_util.h"
 #include "webrtc/test/fake_network_pipe.h"
+#include "webrtc/test/gmock.h"
+#include "webrtc/test/gtest.h"
 
 using ::testing::_;
 using ::testing::AnyNumber;
@@ -72,7 +71,7 @@ class FakeNetworkPipeTest : public ::testing::Test {
   }
 
   void SendPackets(FakeNetworkPipe* pipe, int number_packets, int packet_size) {
-    RTC_DCHECK_GE(packet_size, static_cast<int>(sizeof(int)));
+    RTC_DCHECK_GE(packet_size, sizeof(int));
     std::unique_ptr<uint8_t[]> packet(new uint8_t[packet_size]);
     for (int i = 0; i < number_packets; ++i) {
       // Set a sequence number for the packets by
